agents4sci_rej/submission_345/345_Multi_Agent_AI_System_for__Supplementary Material/requirements.txt
# Core Data Science Stack
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0

# Machine Learning (Optional)
xgboost>=1.7.0
shap>=0.42.0

# Web Framework & UI
streamlit>=1.28.0
jinja2>=3.1.0

# Document Processing
pyyaml>=6.0
python-pptx>=0.6.0
fpdf2>=2.7.0
reportlab>=4.0.0

# Testing & Development
pytest>=7.4.0
jupyter>=1.0.0
notebook>=7.0.0

# Multi-LLM Integration (AI Scientist Core)
openai>=1.50.0,<2.0.0
google-generativeai>=0.8.0,<1.0.0  
anthropic>=0.40.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# LangChain Integration (Agent System)
langgraph>=0.0.40
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0,<3.0.0

# Web Scraping & Reference Validation
requests>=2.31.0
beautifulsoup4>=4.12.0
scholarly>=1.7.0

# Statistical Analysis Extensions (AI Scientist)
plotly>=5.15.0
bokeh>=3.2.0
tabulate>=0.9.0

# Academic Paper Generation (Results Analyst)
latex>=0.7.0
pypdf>=3.15.0
python-docx>=0.8.11

# Data Export & Compliance (Compliance Manager)
openpyxl>=3.1.0
xlsxwriter>=3.1.0
jsonschema>=4.19.0
